backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
earthengine-api==0.1.412
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
rasterio==1.4.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.28.1
reportlab==4.2.5
diskcache==5.6.3
python-dotenv==1.0.1
shapely==2.0.6
geojson==3.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
